update 13 april 11:16

+++ routes.py
@@ -1,10 +1,9 @@
 from flask import Flask,render_template,redirect,url_for,flash,request,session
 from app import app
 from functools import wraps
-from models import db, User,Book,BookRequest,Feedback,AccessLog,Section
+from models import db, User,Book,BookRequest,Feedback,Section
 from werkzeug.security import generate_password_hash ,check_password_hash
 from datetime import datetime ,timedelta
-
 
 
 @app.route('/register')
@@ -19,7 +18,7 @@ def register_post():
     name = request.form.get('name')
     is_librarian = int(request.form.get('userType'))
 
-    if not username or not password or not confirm_password:
+    if not username or not password or not confirm_password or not name:
         flash('Please fill out all fields')
         return redirect(url_for('register'))
 
@@ -41,7 +40,12 @@ def register_post():
 ##################################################
 @app.route('/login')
 def login():
-    return render_template('login.html')
+    try:
+        user = User.query.get(session['user_id'])
+        flash('you are logged in')
+        return redirect(url_for('index',user=user))
+    except:
+        return render_template('login.html')
 
 @app.route('/login',methods=['POST'])
 def login_post():
@@ -94,14 +98,11 @@ def librarian_required(func):
 ############################################
 
 
-
-
 @app.route('/profile')
 @auth_required
 def profile():
     user = User.query.get(session['user_id'])
     return render_template('profile.html', user=user)
-    
 
 @app.route('/profile',methods=['POST'])
 @auth_required
@@ -146,35 +147,10 @@ def librarian():
     sections = Section.query.all()
     section_names = [section.name for section in sections]
     return render_template('librarian.html', sections=sections, section_names=section_names,user= user )
-#########################
-
-@app.route('/')
-@auth_required
-def index():
-    user = User.query.get(session['user_id'])
-    if user.is_librarian:
-        return redirect(url_for('librarian'))
-    
-    sections = Section.query.all()
-    for section in sections:
-        for book in section.books:
-            ratings= Feedback.query.filter_by(book_id=book.id).all()
-            if ratings:
-                valid_ratings = [rating.rating for rating in ratings if rating.rating is not None]
-                avg_rating = sum(valid_ratings)/len(valid_ratings)
-                book.book_rating = avg_rating
 
 
-    sname = request.args.get('sname') or ''
-    bname = request.args.get('bname') or ''
-    aname = request.args.get('aname')
-
-    if sname:
-        sections = Section.query.filter(Section.name.ilike(f'%{sname}%')).all()
 
 
-    db.session.commit()
-    return render_template('index.html',user = user, sections=sections, bname=bname, aname=aname, sname= sname)
 
 ################################
 
@@ -184,14 +160,13 @@ def add_section():
     user = User.query.get(session['user_id'])
     return render_template('section/add.html',user=user)
 
-
 @app.route('/section/add', methods=['POST'])
 @librarian_required
 def add_section_post():
     name = request.form.get('name')
     description = request.form.get('description')
 
-    if not name:
+    if not name or not description:
         flash('Please fill out all fields')
         return redirect(url_for('add_section'))
     
@@ -201,7 +176,7 @@ def add_section_post():
 
     flash('Section added successfully')
     return redirect(url_for('librarian'))
-    
+##################################
 
 @app.route('/section/<int:id>/edit')
 @librarian_required
@@ -219,6 +194,10 @@ def edit_section_post(id):
     name = request.form.get('name')
     description= request.form.get('description')
 
+    if not name or not description:
+        flash('Please fill out all fields')
+        return redirect(url_for('edit_section',id=section.id))
+    
     section.name = name
     section.description = description
     db.session.commit()
@@ -250,7 +229,7 @@ def delete_section_post(id):
     flash('Section deleted successfully')
     return redirect(url_for('librarian'))
 
-#######################################################
+####################################################
 
 @app.route('/book/add/<int:section_id>')
 @librarian_required
@@ -259,7 +238,6 @@ def add_book(section_id):
     sections = Section.query.all()
     section = Section.query.get(section_id)
     return render_template('book/add.html', section=section, sections=sections,user=user)
-
 
 @app.route('/book/add/', methods=['POST'])
 @librarian_required
@@ -270,12 +248,15 @@ def add_book_post():
     section_id = request.form.get('section_id')
     section = Section.query.get(section_id)
 
+    if not name or not authors or not content :
+        flash('Please fill out all fields')
+        return redirect(url_for('add_book',section_id=section.id))
+    
     book = Book(name=name, authors=authors, section=section, content=content,section_id=section_id)
     db.session.add(book)
     db.session.commit()
     flash('Product added successfully')
     return redirect(url_for('show_section', id=section_id))
-
 
 @app.route('/book/<int:id>/show')
 @librarian_required
@@ -283,9 +264,6 @@ def show_book(id):
     user = User.query.get(session['user_id'])
     book = Book.query.get(id)
     return render_template('book/show.html',book= book,user=user)
-
-
-
 
 @app.route('/book/<int:id>/edit')
 @librarian_required
@@ -302,7 +280,9 @@ def edit_book_post(id):
     section_id = request.form.get('section_id')
     content = request.form.get('content')
     authors = request.form.get('authors')
-
+    if not name or not section_id or not content or not authors:
+        flash('Please fill out all fields')
+        return redirect(url_for('edit_book',id=id))
     section = Section.query.get(section_id)
 
     book = Book.query.get(id)
@@ -341,8 +321,33 @@ def delete_book_post(id):
     flash('Book deleted successfully')
     return redirect(url_for('show_section', id=section_id))
 
-###########################
 
+#########################
+@app.route('/')
+@auth_required
+def index():
+    user = User.query.get(session['user_id'])
+    if user.is_librarian:
+        return redirect(url_for('librarian'))
+    
+    sections = Section.query.all()
+    for section in sections:
+        for book in section.books:
+            feedback= Feedback.query.filter_by(book_id=book.id).all()
+            if feedback:
+                valid_ratings = [feedback.rating for feedback in feedback if feedback.rating is not None]
+                avg_rating = sum(valid_ratings)/len(valid_ratings)
+                book.book_rating = avg_rating
+
+    sname = request.args.get('sname') or ''
+    bname = request.args.get('bname') or ''
+    aname = request.args.get('aname')
+    if sname:
+        sections = Section.query.filter(Section.name.ilike(f'%{sname}%')).all()
+    db.session.commit()
+    return render_template('index.html',user = user, sections=sections, bname=bname, aname=aname, sname= sname)
+
+###########################
 @app.route('/mybook')
 @auth_required
 def mybook():
@@ -361,7 +366,6 @@ def mybook_read(book_request_id):
             book_request.access_expiry = book_request.issued_date + timedelta(days=book_request.requested_day)
 
     if book_request.access_expiry < datetime.utcnow() :
-        book_request.status = "Expired"
         db.session.delete(book_request)
         db.session.commit()  
     
@@ -379,37 +383,26 @@ def mybook_read(book_request_id):
 def myrequest_return(book_request_id):
     mybook_request = BookRequest.query.get(book_request_id)
     if mybook_request and mybook_request.status == 'Pending':
-        mybook_request.status = 'Rejected'
-        if mybook_request.status== 'Rejected':
-            db.session.delete(mybook_request)
+        db.session.delete(mybook_request)
         db.session.commit()
-        flash(' Book request return ')
+        flash(' Book Request Return ')
     else:
-        flash('Invalid book return request')
+        flash('Invalid Book Return Request')
     return redirect(url_for('mybook'))  
 
 
 @app.route('/mybook_return/<int:book_id>',methods=['POST'])
 @auth_required
-def myreqeust_return(book_id):
+def mybook_return(book_id):
     book = BookRequest.query.get(book_id)
+    if not book:
+        flash('Invalid Book Return Request')
+        return redirect(url_for('mybook'))
     db.session.delete(book)
     db.session.commit()
-    flash('book Returnd successfully')
+    flash('Book Returnd successfully')
     return redirect(url_for('mybook'))
 
-
-
-
-
-
-@app.route('/book_request')
-@auth_required
-def book_request():
-    user = User.query.get(session['user_id'])
-    request = BookRequest.query.filter_by(user_id=session['user_id']).all()
-    
-    return render_template('requests/send_req.html',request= request,user=user)
 
 
 @app.route('/book_request,<int:book_id>', methods=['POST'])
@@ -418,8 +411,8 @@ def book_request_post(book_id):
     days =  int(request.form.get('requested_day'))
     user_id = session['user_id']
     
-    if days > 30 and days < 0 :
-        flash('You can request max 20 days')
+    if days > 30 or days < 0 :
+        flash('You can request max 30 days')
         return redirect(url_for('index'))
 
     check_exist = BookRequest.query.filter_by(user_id=user_id, book_id=book_id).first()
@@ -428,10 +421,10 @@ def book_request_post(book_id):
         flash('You have already requested for this book')
         return redirect(url_for('index'))
 
-    num_of_requests = BookRequest.query.filter_by(user_id=user_id, status='Pending').count()
+    num_of_request = BookRequest.query.filter_by(user_id=user_id, status='Pending').count()
 
-    if  num_of_requests >= 5:
-        flash('You have reached the maximum limit of 5 requests')
+    if  num_of_request >= 5:
+        flash('You have reached the maximum limit of 5 Book Requests')
         return redirect(url_for('index'))
 
     
@@ -459,7 +452,7 @@ def process_request(request_id, action):
     book_request = BookRequest.query.get(request_id)
 
     if not book_request:
-        flash('Book request not found.')  
+        flash('Book request not found')  
         return redirect(url_for('see_requests'))
     if action == 'grant':
         book_request.issued_date = datetime.utcnow()
@@ -469,9 +462,6 @@ def process_request(request_id, action):
        
        
     elif action == 'reject':
-        book_request.status = 'Rejected'
-        book_request.issued_date = None
-        book_request.access_expiry = None
         db.session.delete(book_request)
 
     db.session.commit()
@@ -487,6 +477,10 @@ def process_request(request_id, action):
 def myfeedback(book_id):
     user = User.query.get(session['user_id'])
     book_request = BookRequest.query.get(book_id)
+    if not book_request:
+        flash('Book Reqeust Not Found')
+        return redirect(url_for('mybook'))
+
     return render_template('feedback.html', book_request=book_request,user=user)
 
 
@@ -496,43 +490,37 @@ def myfeedback_post(book_id):
     book_request = BookRequest.query.get(book_id)
 
     if not book_request:
-        flash('Book request not found')
+        flash('Book Reqeust Not Found')
         return redirect(url_for('mybook'))
 
-    if request.method == 'POST':
-        rating = request.form.get('rating')
-        feedback_text = request.form.get('feedback')
-
-        feedback = Feedback(user_id=session['user_id'],book_id=book_request.book_id,rating=rating,feedback=feedback_text)
-        db.session.add(feedback)
-        db.session.commit()
-
-        flash('Feedback sent successfully.')
+    check_feedback= Feedback.query.filter_by(user_id = session['user_id'],book_id = book_request.book_id).first()
+    if check_feedback:
+        flash('You have already submitted feedback for this book ')
         return redirect(url_for('mybook'))
 
-    return render_template('feedback.html', book_request=book_request)
+    rating = request.form.get('rating')
+    feedback_text = request.form.get('feedback')
 
-
+    feedback = Feedback(user_id=session['user_id'],book_id=book_request.book_id,rating=rating,feedback=feedback_text)
+    
+    db.session.add(feedback)
+    db.session.commit()
+    
+    flash('Feedback  submitted successfully.')
+    return redirect(url_for('mybook'))
 
 #######################################################
-
-
-
-
-
-
-
 
 @app.route('/book_detail/<int:book_id>')
 @auth_required
 def book_detail(book_id):
     user = User.query.get(session['user_id'])
-    book = Book.query.get_or_404(book_id)
-    ratings = Feedback.query.filter_by(book_id=book.id).all()
-    valid_ratings = [rating.rating for rating in ratings if rating.rating is not None]
-    avg_rating = sum(valid_ratings) / len(valid_ratings) if valid_ratings else 0
-
-    return render_template('book_detail.html', user=user , book=book, avg_rating=avg_rating, ratings=ratings)
+    book = Book.query.get(book_id)
+    feedbacks = Feedback.query.filter_by(book_id=book.id).all()
+    if not book:
+        flash('Book does not exist')
+        return redirect(url_for('mybook'))
+    return render_template('book_detail.html', user=user , book=book,feedbacks=feedbacks)
